Slight change at line 1017 to avoid unsigned/signed warning.
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@8803 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/iem/comport/comport/comport.c
@@ -1014,7 +1014,7 @@ static int write_serial(t_comport *x, unsigned char  serial_byte)
 static int write_serials(t_comport *x, unsigned char *serial_buf, size_t buf_length)
 {
     int result = write(x->comhandle,(char *)serial_buf, buf_length);
-    if (result != buf_length)
+    if (result != (int)buf_length)
         post ("[comport] write returned %d, errno is %d", result, errno);
     return result;
 }
